1V without fact

+++ fullcars-restapi/src/main/java/com/fullcars/restapi/dto/SalesData.java
@@ -1,16 +1,17 @@
 package com.fullcars.restapi.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class SalesData {
-    private final long amount;
+    private final BigDecimal amount;
     private final LocalDate date;
 
-    public SalesData(long amount, LocalDate date) {
+    public SalesData(BigDecimal amount, LocalDate date) {
         this.amount = amount;
         this.date = date;
     }
 
-    public long getAmount() { return amount; }
+    public BigDecimal getAmount() { return amount; }
     public LocalDate getDate() { return date; }
 }
